tranfer the selu_op and pass the CI
